Move auth tests
