+++ packages/core/test/request-builder/header-builder/header-builder.spec.ts
@@ -1,21 +1,20 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
+
 import nock from 'nock';
 import { buildHeaders } from '../../../src/request-builder/header-builder/header-builder';
-import { ODataCreateRequestConfig } from '../../../src/request-builder/request/odata-create-request-config';
-import { ODataGetAllRequestConfig } from '../../../src/request-builder/request/odata-get-all-request-config';
-import { ODataRequest } from '../../../src/request-builder/request/odata-request';
-import { ODataUpdateRequestConfig } from '../../../src/request-builder/request/odata-update-request-config';
-import { sanitizeDestination } from '../../../src/scp-cf';
 import { Destination } from '../../../src/scp-cf/destination-service-types';
 import { mockedConnectivityServiceProxyConfig } from '../../test-util/environment-mocks';
 import { muteLoggers } from '../../test-util/mute-logger';
 import {
-  defaultBasicCredentials,
   defaultDestination,
   mockHeaderRequest
 } from '../../test-util/request-mocker';
-import { TestEntity } from '../../test-util/test-services/test-service';
 import * as csrfHeaders from '../../../src/request-builder/header-builder/csrf-token-header';
+import {
+  createGetAllRequest,
+  createUpdateRequest,
+  createCreateRequest
+} from '../../test-util/create-requests';
 
 describe('Header-builder:', () => {
   beforeAll(() => {
@@ -29,20 +28,6 @@ describe('Header-builder:', () => {
 
     const headers = await buildHeaders(request);
     expect(headers.authorization).toBe(authString);
-  });
-
-  it('creates basic authentication headers from destination credentials.', async () => {
-    const request = createGetAllRequest(defaultDestination);
-
-    const headers = await buildHeaders(request);
-    expect(headers.authorization).toBe(defaultBasicCredentials);
-  });
-
-  it('creates no authentication headers when only url is defined in a destination.', async () => {
-    const request = createGetAllRequest({ url: defaultDestination.url });
-
-    const headers = await buildHeaders(request);
-    expect(headers.authorization).toBeUndefined();
   });
 
   it('uses authTokens if present on a destination', async () => {
@@ -62,76 +47,6 @@ describe('Header-builder:', () => {
     const headers = await buildHeaders(request);
 
     expect(headers.authorization).toBe('Bearer some.token');
-  });
-
-  it('throws an error if the error property is truthy for all provided authTokens', async () => {
-    const destination: Destination = {
-      ...defaultDestination,
-      authentication: 'OAuth2SAMLBearerAssertion',
-      authTokens: [
-        {
-          type: 'Bearer',
-          value: 'some.token',
-          expiresIn: '3600',
-          error: 'error'
-        },
-        {
-          type: 'Bearer',
-          value: 'some.other.token',
-          expiresIn: '3600',
-          error: 'error'
-        }
-      ]
-    };
-    const request = createGetAllRequest(destination);
-
-    await expect(buildHeaders(request)).rejects.toThrowErrorMatchingSnapshot();
-  });
-
-  it('uses the first authToken where the error property is falsy', async () => {
-    const destination: Destination = {
-      ...defaultDestination,
-      authentication: 'OAuth2SAMLBearerAssertion',
-      authTokens: [
-        {
-          type: 'Bearer',
-          value: 'some.token',
-          expiresIn: '3600',
-          error: 'error'
-        },
-        {
-          type: 'Bearer',
-          value: 'some.other.token',
-          expiresIn: '3600',
-          error: null
-        }
-      ]
-    };
-    const request = createGetAllRequest(destination);
-    const headers = await buildHeaders(request);
-    expect(headers.authorization).toBe('Bearer some.other.token');
-  });
-
-  it('Throws an error if no authTokens are present, username is null and authenticationType is wrongly set', async () => {
-    const destination: Destination = {
-      ...defaultDestination,
-      username: null,
-      authentication: 'BasicAuthentication'
-    };
-    const request = createGetAllRequest(destination);
-
-    await expect(buildHeaders(request)).rejects.toThrowErrorMatchingSnapshot();
-  });
-
-  it('Throws an error if no authTokens are present, password is null and authenticationType is wrongly set', async () => {
-    const destination: Destination = {
-      ...defaultDestination,
-      password: null,
-      authentication: 'BasicAuthentication'
-    };
-    const request = createGetAllRequest(destination);
-
-    await expect(buildHeaders(request)).rejects.toThrowErrorMatchingSnapshot();
   });
 
   describe('update request header with etag', () => {
@@ -237,24 +152,24 @@ describe('Header-builder:', () => {
   });
 
   it('Skips csrf token retrieval for existing csrf header', async () => {
-    spyOn(csrfHeaders, 'buildCsrfHeaders2');
+    spyOn(csrfHeaders, 'buildCsrfHeaders');
     const request = createCreateRequest(defaultDestination);
     request.config.customHeaders = { 'x-csrf-token': 'defined' };
 
     mockHeaderRequest({ request });
 
     await request.headers();
-    expect(csrfHeaders.buildCsrfHeaders2).not.toHaveBeenCalled();
+    expect(csrfHeaders.buildCsrfHeaders).not.toHaveBeenCalled();
   });
 
   it('Skips csrf token retrieval for GET request', async () => {
-    spyOn(csrfHeaders, 'buildCsrfHeaders2');
+    spyOn(csrfHeaders, 'buildCsrfHeaders');
     const request = createGetAllRequest(defaultDestination);
 
     mockHeaderRequest({ request });
 
     await request.headers();
-    expect(csrfHeaders.buildCsrfHeaders2).not.toHaveBeenCalled();
+    expect(csrfHeaders.buildCsrfHeaders).not.toHaveBeenCalled();
   });
 
   it('Prioritizes custom Authorization headers (upper case A)', async () => {
@@ -276,109 +191,4 @@ describe('Header-builder:', () => {
     const headers = await request.headers();
     expect(headers.authorization).toBe('Basic SOMETHINGSOMETHING');
   });
-
-  describe('OAuth2ClientCredentials', () => {
-    const destination: Destination = {
-      ...defaultDestination,
-      authentication: 'OAuth2ClientCredentials',
-      tokenServiceUrl: 'https://token.example.com/oauth/token',
-      clientId: 'TokenClientId',
-      clientSecret: 'TokenClientSecret'
-    };
-
-    const requestBody =
-      'grant_type=client_credentials&client_id=TokenClientId&client_secret=TokenClientSecret';
-    const fakeToken = 'FakeOAuth2ClientCredentialsToken';
-
-    it('should add access token to the request header for "OAuth2ClientCredentials" authentication', async () => {
-      nock('https://token.example.com', {
-        reqheaders: {
-          'content-type': 'application/x-www-form-urlencoded',
-          authorization: 'Basic VG9rZW5DbGllbnRJZDpUb2tlbkNsaWVudFNlY3JldA=='
-        }
-      })
-        .post('/oauth/token', requestBody)
-        .once()
-        .reply(200, { access_token: fakeToken, expires_in: 0, scope: '' });
-
-      const request = createGetAllRequest(destination);
-      const actual = await buildHeaders(request);
-
-      expect(actual).toEqual({
-        authorization: 'Bearer ' + fakeToken,
-        'content-type': 'application/json',
-        accept: 'application/json',
-        'sap-client': '123'
-      });
-    });
-
-    it('should use  tokenServiceUser and tokenServicePassword in header authorization when specified', async () => {
-      destination.tokenServiceUser = 'TokenServiceUser';
-      destination.tokenServicePassword = 'TokenServicePassword';
-      nock('https://token.example.com', {
-        reqheaders: {
-          'content-type': 'application/x-www-form-urlencoded',
-          authorization:
-            'Basic VG9rZW5TZXJ2aWNlVXNlcjpUb2tlblNlcnZpY2VQYXNzd29yZA=='
-        }
-      })
-        .post('/oauth/token', requestBody)
-        .once()
-        .reply(200, { access_token: fakeToken, expires_in: 0, scope: '' });
-
-      const request = createGetAllRequest(destination);
-      const actual = await buildHeaders(request);
-
-      expect(actual).toEqual({
-        authorization: 'Bearer ' + fakeToken,
-        'content-type': 'application/json',
-        accept: 'application/json',
-        'sap-client': '123'
-      });
-    });
-
-    it('should throw an error if tokenServiceUrl, client_id or client_secret are not set', async () => {
-      delete destination.tokenServiceUrl;
-
-      const request = createGetAllRequest(destination);
-      await expect(
-        buildHeaders(request)
-      ).rejects.toThrowErrorMatchingSnapshot();
-    });
-
-    it('should throw an error when the access token to the request header for "OAuth2ClientCredentials" is rejected', async () => {
-      nock('https://token.example.com', {
-        reqheaders: {
-          'content-type': 'application/x-www-form-urlencoded',
-          authorization: 'Basic VG9rZW5DbGllbnRJZDpUb2tlbkNsaWVudFNlY3JldA=='
-        }
-      })
-        .post('/oauth/token', requestBody)
-        .once()
-        .reply(401, {
-          error: 'unauthorized',
-          error_description: 'Bad credentials'
-        });
-
-      const request = createGetAllRequest(destination);
-      await expect(
-        buildHeaders(request)
-      ).rejects.toThrowErrorMatchingSnapshot();
-    });
-  });
 });
-
-function createUpdateRequest(dest: Destination) {
-  const requestConfig = new ODataUpdateRequestConfig(TestEntity);
-  return new ODataRequest(requestConfig, sanitizeDestination(dest));
-}
-
-function createGetAllRequest(dest: Destination) {
-  const requestConfig = new ODataGetAllRequestConfig(TestEntity);
-  return new ODataRequest(requestConfig, sanitizeDestination(dest));
-}
-
-function createCreateRequest(dest: Destination) {
-  const requestConfig = new ODataCreateRequestConfig(TestEntity);
-  return new ODataRequest(requestConfig, sanitizeDestination(dest));
-}
